Update Player.js
The player character rotates but I still ain't happy with the results. Gonna try making a function that simulates the Bezier curve but by making use of the de Casteljau algorithm.

I feel I might be doing way too much for what should be a simple thing to get working.