Update index.js
Added orbit controls and some new navmeshes for different arrangements of the test scene.

Buggy as hell. I'm gonna have to make it on my own using Blender for the sake of testing the pathfinding. The pathfinding is also buggy and is likely because of making use of primitive shapes. Who knows?