Update index.js
Improved player movement along the navpath immensely!

All nodes in the navpath, except for the last one, will be raised by a certain amount of units and that dramatically changes the way the player object traverses. Insane stuff.

It's still 90% perfect. The player sometimes gets stuck with the world geometry while cornering for some reason. Need to rage test along with / without the "resolve nodes" code and then with slopes.