Update index.js
Adding interpolation to fix the lack of nodes that appear when traversing slopes.

Sometimes, they appear but its like 1 in 20 cases. And the only solution I can think of is interpolating the initially created navpath with more nodes and then adjusting their heights accordingly.

Buggy as hell. Never expected this one to take this much time.