Update index.js
Lots of stuff. For starters, I tried making my own version of Unity's Physics.CheckSphere for improving the adjustNodePosition() code because I felt that passing the player object in it as well just made no sense. Tried excluding it and voila, everything went haywire.

The navpath is generating nicely along the slopes but in the level, it is sometimes behaving rather weirdly and is going off kilter here and there for some weird reason. And when any other area is clicked lateral to the upper slope exit, the generated navpath looks really weird and it will result in the player flying yet again.

I don't know. I feel I'm going wrong somewhere if I'm working this hard.